Overhaul the backup, restore, and upgrade processes

+++ scripts/app-upgrade.sh
@@ -3,17 +3,23 @@ set -e
 
 START_TIME=$(date +%s)
 
+if [ ! -d .git ]; then
+    echo "[E] This script needs to run from the top directory of the repo. Current working directory:"
+    echo "      $(pwd)"
+    exit 1
+fi
+
 echo "=== Shutting down web container. ==============================================="
 docker-compose stop web
 
 echo "=== Shutting down sync container. =============================================="
 docker-compose stop sync
 
-echo "=== Starting backup container. ================================================="
-docker-compose up -d backup
+echo "=== Starting cron container. ==================================================="
+docker-compose up -d cron
 
 echo "=== Backing up application stack. =============================================="
-docker-compose exec backup app-backup
+docker-compose exec cron app-backup
 
 echo "=== Removing currnet application stack. ========================================"
 docker-compose down
@@ -31,22 +37,19 @@ echo "[I] Upgrading Margarita from '$OLD_MARGARITA_VERSION' to '$NEW_MARGARITA_V
 sed -i.bak -e "s/^REPOSADO_VERSION=.*/REPOSADO_VERSION=$NEW_REPOSADO_VERSION/g" -e "s/^MARGARITA_VERSION=.*/MARGARITA_VERSION=$NEW_MARGARITA_VERSION/g" .env
 
 echo "=== Deleting old images. ======================================================="
-IMAGE_BACKUP=$(docker images ianharrier/applesus-backup -q)
+IMAGE_CRON=$(docker images ianharrier/applesus-cron -q)
 IMAGE_SYNC=$(docker images ianharrier/reposado -q)
 IMAGE_WEB=$(docker images ianharrier/margarita -q)
-docker rmi $IMAGE_BACKUP $IMAGE_SYNC $IMAGE_WEB
+docker rmi $IMAGE_CRON $IMAGE_SYNC $IMAGE_WEB
 
 echo "=== Building new images. ======================================================="
-docker-compose build --pull
-
-echo "=== Starting backup container. ================================================="
-docker-compose up -d backup
+docker-compose build --pull --no-cache
 
 echo "=== Restoring application stack to most recent backup. ========================="
 cd backups
 LATEST_BACKUP=$(ls -1tr *.tar.gz 2> /dev/null | tail -n 1)
 cd ..
-docker-compose exec backup app-restore $LATEST_BACKUP
+./scripts/app-restore.sh $LATEST_BACKUP
 
 END_TIME=$(date +%s)
